backend/requirements.txt
django
dotenv
django-cors-headers
djangorestframework
djangorestframework-simplejwt
"psycopg[binary]"
